requirements/base.txt
aiocache==0.11.1
aiohttp==3.8.1
aioredis==1.3.1
backoff==1.11.1
elasticsearch[async]==7.15.0
fastapi==0.70.0
ipython==7.28.0
jsonpickle==2.0.0
orjson==3.6.4
python-jose[cryptography]==3.3.0
typer==0.4.0
uvicorn==0.15.0
uvloop==0.16.0